backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
Werkzeug>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-docx>=1.0.0

